requirements.txt
aiogram==3.0.0b7
docker_py==1.10.6
psycopg2==2.9.5
python-dotenv==1.0.0
Requests==2.28.2
